Phase 4: kill duplication, fix ownership & layering (behavior-preserving)
Part of the Fable 5 Findings epic (#193); addresses issue #197. Built on the
integration of Phases 1/2/3/5 so the deduped code keeps those fixes.

Behavior-preserving consolidations (existing tests pass unchanged):
- studyfunc: extract the k-means kernel into _cluster_kmeans.py so pop_clust,
  optimal_kmeans and robust_kmeans import it downward (layering inversion
  fixed); share unique_preserving_order, trialinfo_rows, default_measure_target
  and the MEASURE_DATA_FIELDS map.
- timefreq: newtimef/newcrossf use the canonical statistics.fdr and the shared
  bootstrap/threshold helpers instead of forked copies; _is_on imports the
  canonical whitelist predicate.
- popfunc: remove the duplicated _is_on copies (import _pop_utils.is_on),
  pop_topoplot reuses the shared python_literal, pop_topochansel delegates to
  pop_chansel, pop_comperp collapses the repeated lowpass calls, pop_saveset
  serializes chanlocs through the single _chanlocs_to_struct_array, and the
  near-identical pop_rejkurt/pop_jointprob share their rejection plumbing.
- gui/cli/extensions: menu_actions collapses the per-name dispatch branches;
  the pop-result interpretation contract is shared; the dead run_main() is
  removed; extension_catalog splits its manager and curation halves
  (extension_catalog_validation.py) and reuses the canonical entry-point
  helpers; plugin_menu derives bundled metadata from the registry.
- plugins: firws/firwsord move to the firfilt plugin (correct ownership) with
  clean_rawdata importing them.

Findings that would change observable behavior (figure titles, history
strings, cluster partition/provenance, error strings) or need files outside
this scope are intentionally left for a follow-up, as are the four large-file
decompositions (pop_load_frombids, statistics/_core, pop_prop_extended,
QtDialogRenderer) — those are tracked as Phase 4b.
