+++ src/eegprep/cli/core.py
@@ -291,44 +291,3 @@ def _input_file_record(path: Path | dict[str, Any]) -> dict[str, Any]:
     if isinstance(path, dict):
         return json_safe(path)
     return {"path": str(path), "sha256": sha256_file(path)}
-
-
-def run_main(handler: Any, args: Any) -> int:
-    try:
-        result = handler(args)
-        print_result(result, as_json=bool(getattr(args, "json", False)))
-        return 1 if result.get("status") == "error" else 0
-    except EEGPrepCLIError as exc:
-        print_result(exc.to_response(), as_json=bool(getattr(args, "json", False)))
-        return exc.exit_code
-    except Exception as exc:
-        code = getattr(exc, "code", None)
-        message = getattr(exc, "message", None)
-        if code and message:
-            payload = {
-                "status": "error",
-                "schema_version": "eegprep.error.v1",
-                "code": code,
-                "message": message,
-            }
-            path = getattr(exc, "path", None)
-            suggestion = getattr(exc, "suggestion", None)
-            if path is not None:
-                payload["path"] = str(path)
-            if suggestion is not None:
-                payload["suggestion"] = suggestion
-            print_result(payload, as_json=bool(getattr(args, "json", False)))
-            return int(getattr(exc, "exit_code", 1) or 1)
-        error = EEGPrepCLIError(
-            "UNEXPECTED_ERROR",
-            str(exc),
-            suggestion="Rerun with --verbose or file an issue if this is reproducible.",
-        )
-        print_result(error.to_response(), as_json=bool(getattr(args, "json", False)))
-        if bool(getattr(args, "verbose", False)):
-            raise
-        return 1
-
-
-def eprint(message: str) -> None:
-    print(message, file=sys.stderr)
