+++ src/eegprep/functions/popfunc/eeg_picard.py
@@ -4,7 +4,7 @@ import copy
 
 from picard import picard
 import numpy as np
-from ._ica_utils import flatten_ica_data, reshape_ica_activations
+from ._ica_utils import finalize_ica_fields, flatten_ica_data, reshape_ica_activations
 from ..miscfunc.pinv import pinv
 
 
@@ -77,36 +77,4 @@ def eeg_picard(EEG, engine=None, posact='off', sortcomps='off', **kwargs):
         # sorting/normalization options)
         EEG = engine.eeg_picard(EEG, **kwargs)
 
-    # optionally sort components by mean descending activation variance
-    if sortcomps in ('on', True):
-        # Flatten icaact to 2D for variance computation
-        icaact_2d = flatten_ica_data(EEG['icaact'])
-        # Compute variance metric: sum(icawinv^2) .* sum(icaact^2)
-        variance_metric = np.sum(EEG['icawinv'] ** 2, axis=0) * np.sum(icaact_2d**2, axis=1)
-        # Sort indices in descending order
-        windex = np.argsort(variance_metric)[::-1]
-        # Reorder components
-        EEG['icaact'] = EEG['icaact'][windex, :, :]
-        EEG['icaweights'] = EEG['icaweights'][windex, :]
-        EEG['icawinv'] = EEG['icawinv'][:, windex]
-
-    # optionally normalize components using the same rule as runica()
-    if posact in ('on', True):
-        # Flatten icaact to 2D for finding max abs values
-        icaact_2d = flatten_ica_data(EEG['icaact'])
-        # Find indices of max absolute values for each component
-        ix = np.argmax(np.abs(icaact_2d), axis=1)
-        ncomps = EEG['icaact'].shape[0]
-
-        for r in range(ncomps):
-            if np.sign(icaact_2d[r, ix[r]]) < 0:
-                # A sign flip commutes through the factorization, so negate the
-                # matching row of icaweights and column of icawinv directly. This
-                # preserves the invariants icawinv == pinv(icaweights @ icasphere)
-                # and icaact == icaweights @ icasphere @ data, leaving icasphere
-                # untouched.
-                EEG['icaact'][r, :, :] = -EEG['icaact'][r, :, :]
-                EEG['icawinv'][:, r] = -EEG['icawinv'][:, r]
-                EEG['icaweights'][r, :] = -EEG['icaweights'][r, :]
-
-    return EEG
+    return finalize_ica_fields(EEG, sortcomps=sortcomps, posact=posact)
