+++ src/eegprep/functions/studyfunc/pop_clust.py
@@ -9,14 +9,12 @@ import numpy as np
 from eegprep.functions.guifunc.inputgui import inputgui
 from eegprep.functions.guifunc.spec import ControlSpec, DialogSpec
 from eegprep.functions.popfunc._plot_utils import numeric_vector
+from eegprep.functions.studyfunc._cluster_kmeans import kmeans_labels
 from eegprep.functions.studyfunc._cluster_utils import checked_study_and_datasets, cluster_command
 from eegprep.functions.studyfunc.std_createclust import std_createclust
 
 
 ALGORITHMS = ("kmeans", "kmeanscluster")
-KMEANS_MAX_ITER = 300
-KMEANS_N_INIT = 10
-KMEANS_TOLERANCE = 1e-8
 
 
 def pop_clust(
@@ -66,7 +64,7 @@ def pop_clust(
     if clus_num > data.shape[0]:
         raise ValueError("Number of clusters cannot exceed the number of preclustered components")
 
-    labels, centers = _kmeans_labels(data, clus_num, random_state)
+    labels, centers = kmeans_labels(data, clus_num, random_state)
     if np.isfinite(outliers):
         if outliers <= 0:
             raise ValueError("Outlier threshold must be greater than 0")
@@ -130,47 +128,6 @@ def pop_clust_dialog_spec(STUDY: dict[str, Any]) -> DialogSpec:
             }
         """,
     )
-
-
-def _kmeans_labels(data: np.ndarray, clus_num: int, random_state: int) -> tuple[np.ndarray, np.ndarray]:
-    rng = np.random.default_rng(random_state)
-    best_labels: np.ndarray | None = None
-    best_centers: np.ndarray | None = None
-    best_inertia = float("inf")
-    for _attempt in range(KMEANS_N_INIT):
-        centers = data[rng.choice(data.shape[0], size=clus_num, replace=False)].copy()
-        labels = np.zeros(data.shape[0], dtype=int)
-        for _iteration in range(KMEANS_MAX_ITER):
-            labels = np.argmin(_squared_distances(data, centers), axis=1)
-            new_centers = _recompute_centers(data, labels, centers, clus_num)
-            if np.allclose(new_centers, centers, rtol=0, atol=KMEANS_TOLERANCE):
-                centers = new_centers
-                break
-            centers = new_centers
-        distances = _squared_distances(data, centers)
-        inertia = float(np.sum(distances[np.arange(data.shape[0]), labels]))
-        if inertia < best_inertia:
-            best_inertia = inertia
-            best_labels = labels.copy()
-            best_centers = centers.copy()
-    if best_labels is None or best_centers is None:
-        raise ValueError("K-means failed to initialize clusters")
-    return best_labels.astype(int) + 1, best_centers
-
-
-def _squared_distances(data: np.ndarray, centers: np.ndarray) -> np.ndarray:
-    diff = data[:, np.newaxis, :] - centers[np.newaxis, :, :]
-    return np.sum(diff * diff, axis=2)
-
-
-def _recompute_centers(data: np.ndarray, labels: np.ndarray, centers: np.ndarray, clus_num: int) -> np.ndarray:
-    new_centers = np.empty_like(centers)
-    nearest_distance = np.min(_squared_distances(data, centers), axis=1)
-    fallback_index = int(np.argmax(nearest_distance))
-    for cluster in range(clus_num):
-        rows = data[labels == cluster]
-        new_centers[cluster] = np.mean(rows, axis=0) if rows.size else data[fallback_index]
-    return new_centers
 
 
 def _mark_outliers(data: np.ndarray, labels: np.ndarray, centers: np.ndarray, threshold: float) -> np.ndarray:
